+++ build.gradle.kts
@@ -37,7 +37,7 @@ dependencies {
     implementation("com.zaxxer:HikariCP:5.0.1")
 //    compileOnly("io.th0rgal:oraxen:1.186.0")
     compileOnly("com.nexomc:nexo:1.1.0")
-    implementation("net.dv8tion:JDA:5.2.1")
+    implementation("net.dv8tion:JDA:5.6.1")
     compileOnly("net.citizensnpcs:citizens-main:2.0.37-SNAPSHOT")
     //implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1")
     compileOnly("com.github.LoneDev6:API-ItemsAdder:3.6.3-beta-14")
